+++ src/main/java/com/example/Bar/converter/MenuItemConverter.java
@@ -5,6 +5,8 @@ import com.example.Bar.dto.menuItem.MenuItemDTO;
 import com.example.Bar.entity.MenuItemEntity;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @Component
 public class MenuItemConverter {
 
@@ -14,7 +16,7 @@ public class MenuItemConverter {
         menuItemDTO.setName(menuItemEntity.getName());
         menuItemDTO.setCategory(menuItemEntity.getCategory());
         menuItemDTO.setDescription(menuItemEntity.getDescription());
-        menuItemDTO.setPrice(menuItemEntity.getPrice());
+        menuItemDTO.setPrice(menuItemEntity.getPrice().doubleValue());
 
         return menuItemDTO;
     }
@@ -24,7 +26,7 @@ public class MenuItemConverter {
         menuItemEntity.setName(menuItemDTO.getName());
         menuItemEntity.setCategory(menuItemDTO.getCategory());
         menuItemEntity.setDescription(menuItemDTO.getDescription());
-        menuItemEntity.setPrice(menuItemDTO.getPrice());
+        menuItemEntity.setPrice(BigDecimal.valueOf(menuItemDTO.getPrice()));
 
         return menuItemEntity;
     }
